requirements.txt
APScheduler~=3.7.0
Flask~=2.2.2
Flask-APScheduler~=1.12.4
Flask-Limiter~=2.6.3
Flask-Migrate~=3.1.0
Flask-SQLAlchemy~=2.5.1
httpx~=0.15.5
MarkupSafe~=2.1.1
mysql-connector-python~=8.0.30	
PyMySQL~=1.0.2
requests~=2.28.1
selenium~=3.141.0
SQLAlchemy~=1.3.24
flask-redis
jieba~=0.42.1
alembic~=1.8.1
flask_msearch~=0.2.9.2
email-validator
wrapt~=1.14.1
WTForms~=3.0.1
flask-wtf~=1.1.1
email-validator
Pillow~=9.3.0
redis~=4.3.4
Flask-Caching~=2.0.2
Pillow